Fix classes for Java classpath
It seems that Java and Scala have some caveats for interoperability
with nested objects, so we are promoting the Endpoints to a proper
top-level Scala object, and keeping the Endpoints objects as a helper
to transform JValues back to RestResponses for ElasticSearch.

We might just move that to the esplugin project.